Trait-based refactor for CONNECT_UDP
This does much the same as #3717, but for CONNECT_UDP.
