+++ neqo-http3/tests/connect_udp.rs
@@ -9,9 +9,9 @@
 use http::Uri;
 use neqo_common::{Datagram, Tos, event::Provider as _, header::HeadersExt as _, qinfo};
 use neqo_http3::{
-    ConnectUdpEvent, ConnectUdpRequest, ConnectUdpServerEvent, Error, Http3Client,
-    Http3ClientEvent, Http3Parameters, Http3Server, Http3ServerEvent, Http3State, Priority,
-    SessionAcceptAction,
+    ConnectUdpEvent, Error, Http3Client, Http3ClientEvent, Http3Parameters, Http3Server,
+    Http3ServerEvent, Http3State, Priority, SessionAcceptAction,
+    connect_udp::{ServerEvent, ServerSession},
 };
 use neqo_transport::ConnectionParameters;
 use nss::AuthenticationStatus;
@@ -62,17 +62,15 @@ fn establish_new_session() -> (
     Http3Client,
     Http3Server,
     neqo_http3::StreamId,
-    ConnectUdpRequest,
+    ServerSession,
 ) {
     let (mut client, mut proxy, connect_udp_session_id) = initiate_new_session();
     exchange_packets(&mut client, &mut proxy, false, None);
     let proxy_session = proxy
         .events()
         .find_map(|event| {
-            if let Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::NewSession {
-                session,
-                headers,
-            }) = event
+            if let Http3ServerEvent::ConnectUdp(ServerEvent::NewSession { session, headers }) =
+                event
             {
                 assert_eq!(session.stream_id(), connect_udp_session_id);
 
@@ -108,7 +106,7 @@ fn exchange_packets_through_proxy(
     proxy: &mut Http3Server,
     server: &mut Http3Server,
     connect_udp_session_id: neqo_http3::StreamId,
-    proxy_session: &ConnectUdpRequest,
+    proxy_session: &ServerSession,
 ) {
     qinfo!("Processing client_inner");
     while let Some(dgram) = client_inner.process_output(now()).dgram() {
@@ -135,7 +133,7 @@ fn exchange_packets_through_proxy(
         client_outer.process_multiple_input(dgram.iter_mut(), now());
     }
     let server_dgrams = proxy.events().filter_map(|event| match event {
-        Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::Datagram { datagram, session }) => {
+        Http3ServerEvent::ConnectUdp(ServerEvent::Datagram { datagram, session }) => {
             assert_eq!(session.stream_id(), connect_udp_session_id);
             Some(Datagram::from_bytes(
                 DEFAULT_ADDR,
@@ -208,10 +206,7 @@ fn session_lifecycle(client_closes: bool) {
     let (id, datagram) = proxy
         .events()
         .find_map(|event| {
-            if let Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::Datagram {
-                session,
-                datagram,
-            }) = event
+            if let Http3ServerEvent::ConnectUdp(ServerEvent::Datagram { session, datagram }) = event
             {
                 Some((session.stream_id(), datagram))
             } else {
@@ -256,7 +251,7 @@ fn session_lifecycle(client_closes: bool) {
             .find(|event| {
                 matches!(
                     event,
-                    Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::SessionClosed {
+                    Http3ServerEvent::ConnectUdp(ServerEvent::SessionClosed {
                         session,
                         ..
                     }) if session.stream_id() == session_id
@@ -376,10 +371,7 @@ fn server_datagram_before_accept() {
         let proxy_session = proxy
             .events()
             .find_map(|event| {
-                if let Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::NewSession {
-                    session,
-                    ..
-                }) = event
+                if let Http3ServerEvent::ConnectUdp(ServerEvent::NewSession { session, .. }) = event
                 {
                     Some(session)
                 } else {
@@ -447,10 +439,7 @@ fn server_stream_reset_results_in_client_session_close() {
     let proxy_session = proxy
         .events()
         .find_map(|event| {
-            if let Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::NewSession {
-                session, ..
-            }) = event
-            {
+            if let Http3ServerEvent::ConnectUdp(ServerEvent::NewSession { session, .. }) = event {
                 Some(session)
             } else {
                 None
@@ -536,10 +525,8 @@ fn session_lifecycle_with_http_datagram_capsule() {
     let proxy_session = proxy
         .events()
         .find_map(|event| {
-            if let Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::NewSession {
-                session,
-                headers,
-            }) = event
+            if let Http3ServerEvent::ConnectUdp(ServerEvent::NewSession { session, headers }) =
+                event
             {
                 assert_eq!(session.stream_id(), session_id);
                 assert!(
@@ -580,10 +567,7 @@ fn session_lifecycle_with_http_datagram_capsule() {
     let (id, datagram) = proxy
         .events()
         .find_map(|event| {
-            if let Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::Datagram {
-                session,
-                datagram,
-            }) = event
+            if let Http3ServerEvent::ConnectUdp(ServerEvent::Datagram { session, datagram }) = event
             {
                 Some((session.stream_id(), datagram))
             } else {
@@ -631,9 +615,7 @@ fn session_lifecycle_with_http_datagram_capsule() {
 
     let mut count = 0;
     for event in proxy.events() {
-        if let Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::Datagram { session, datagram }) =
-            event
-        {
+        if let Http3ServerEvent::ConnectUdp(ServerEvent::Datagram { session, datagram }) = event {
             assert_eq!(session.stream_id(), session_id);
             assert_eq!(&datagram.as_ref()[..4], PING);
             count += 1;
@@ -653,7 +635,7 @@ fn session_lifecycle_with_http_datagram_capsule() {
         .find(|event| {
             matches!(
                 event,
-                Http3ServerEvent::ConnectUdp(ConnectUdpServerEvent::SessionClosed {
+                Http3ServerEvent::ConnectUdp(ServerEvent::SessionClosed {
                     session,
                     ..
                 }) if session.stream_id() == session_id
